requirements.txt
git+git://github.com/dlorenc/containerregistry.git@setup
enum34==1.1.6
futures==3.1.1
grpcio==1.4.0
grpcio-tools==1.4.0
httplib2==0.10.3
oauth2client==4.1.2
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
rsa==3.4.2
six==1.10.0
